More cleanup and reorganization
